Update release-notes.html.md.erb
Have started to add intended release notes for v1.1. Still a work in progress but can go to staging and be seen as-is during in-progress updates. Added a what's new section because some of the changes between 1.0 and 1.1 would be considering breaking if 1.0 was GA. I'd like to draw attention to these things that we might not list in v1.1 notes alone, but are still good to know for someone going from 1.0 to 1.1