Update troubleshooting.html.md.erb
We've seen a couple of failures now on SRT when users didn't adjust the default Compute instance count (common when leveraging toolsmiths automated envs); this helps users self-resolve. Submitting as a PR because I'd like a 2nd set of eyes on my added copy to ensure I matched the correct terminology for this PAS tile.