Update configuring-healthwatch.html.md.erb
I think that saying that the "DOMAIN is the domain of your Ops Manager foundation." in the automatic grafana configuration is confusing to the least. I think it would be much more accurate to say that `sys.DOMAIN` is the system domain you have configured in the Domains section in your TAS tile setting. Otherwise, users would think that DOMAIN is the Ops Man FQDN.